+++ src/actors.rs
@@ -1,7 +1,6 @@
 
 pub mod player;
 pub mod circle;
+pub mod triangle;
 // pub mod rect;
-// pub mod ellipse;
-// pub mod triangle;
 // pub mod hexagon;