+++ src/actors/triangle.rs
@@ -1,57 +1,71 @@
+use std::sync::mpsc::Sender;
+
 use async_trait::async_trait;
 use macroquad::prelude::*;
 use macroquad::math::Vec2;
 use macroquad::color::Color;
 
-use crate::{collision_system::collider::Collider, event_system::interface::{Drawable, GameEntity, Moveable, Object, Updatable}};   
+use crate::{collision_system::collider::{CircleCollider, Collider}, event_system::{event::{Event, EventType}, interface::{Drawable, Enemy, GameEntity, Moveable, Object, Publisher, Updatable}}, grid_system::grid::EntityType, renderer::artist::{ConfigType, DrawCall}, utils::machine::{StateMachine, StateType}};   
 
-#[derive(Clone, Copy)]
 pub struct Triangle{
+    //Attributes
     id: u64,
     pos: Vec2,
     size: f32,
     speed: f32,
     color: Color,
-    target: Vec2
+    target: Vec2,
+    //Components
+    sender: Sender<Event>,
+    collider: CircleCollider,
+    machine: StateMachine,
+    //State specifics
+    is_alive: bool,
+    //Emittion
+    emittion_configs: Vec<(StateType, ConfigType)>,
 }
-
-impl Triangle{
-    pub fn new(id: u64, pos: Vec2, size: f32, color: Color, player_pos: Vec2) -> Self{
-        return Triangle {
-            id: id,
-            pos: pos, 
-            size: size, 
-            speed: 100.0,
-            color: color,
-            target: player_pos
-        }
-    }
-
-    pub fn get_id(&self) -> u64{
-        return self.id
-    }
-
-    pub fn get_size(&self) -> f32{
-        return self.size
-    }
-}
-
 
 //========== Triangle interfaces =========
 #[async_trait]
 impl Updatable for Triangle{
-    //Review: Could be quite heavy downcasting for Any
     async fn update(&mut self, delta: f32, mut params: Vec<Box<dyn std::any::Any + Send>>) {
-        if let Some(param_item) = params.pop(){
-            if let Some(player_pos) = param_item.downcast_ref::<Vec2>(){
-                self.target = *player_pos;
-                self.move_to(delta);
+        if self.is_alive{
+            //Update target position
+            let mut overide = None;
+
+            while let Some(param_item) = params.pop(){
+                if let Some(player_pos) = param_item.downcast_ref::<Vec2>(){
+                    self.target = *player_pos;
+                }
+                if let Some(overide_pos) = param_item.downcast_ref::<Option<Vec2>>(){
+                    overide = *overide_pos;
+                }
             }
+
+            //Update based on state machine
+            if let Ok(state) = self.machine.get_state().try_lock(){
+                match *state{
+                    StateType::Idle => {
+                        self.machine.transition(StateType::Moving);
+                    },
+                    StateType::Moving => {
+                        self.move_to(delta, overide);
+                    },
+                    StateType::Hit => {
+                        self.set_alive(false);
+                    },
+                    _ => (), //Unreachable
+                }
+            }
+
+            self.collider.update(self.pos);
+            self.publish(Event::new((self.id, EntityType::Enemy, self.pos), EventType::InsertOrUpdateToGrid)).await
         }
     }
 }
 
 impl Object for Triangle{
+    #[inline(always)]
     fn get_pos(&self) -> Vec2{
         return self.pos
     }
@@ -66,41 +80,135 @@ impl Object for Triangle{
 }
 
 impl Moveable for Triangle{
-    fn move_to(&mut self, delta: f32) -> (f32, f32){
-        let new_pos = self.pos.move_towards(self.target, self.speed * delta);
+    #[inline(always)]
+    fn move_to(&mut self, delta: f32, overide: Option<Vec2>) -> (f32, f32){
+        let mut new_pos = overide.unwrap_or(self.target);
+        
+        new_pos = self.pos.move_towards(new_pos, self.speed * delta);
         self.pos = new_pos;
+
         return self.pos.into()
     }
 }
 
 impl Drawable for Triangle{
-    fn draw(&mut self){
-        draw_circle(self.pos.x, self.pos.y, self.size, self.color);
+    #[inline(always)]
+    fn get_draw_call(&self) -> DrawCall {
+        let top_angle = std::f32::consts::PI / 2.0; // 90 degrees
+        
+        // Calculate the three vertices
+        let v1 = Vec2::new(
+            self.pos.x + self.size * f32::cos(top_angle),
+            self.pos.y - self.size * f32::sin(top_angle) // Negative to point upward
+        );
+        
+        // Bottom right vertex (210 degrees)
+        let v2 = Vec2::new(
+            self.pos.x + self.size * f32::cos(top_angle + 2.0 * std::f32::consts::PI / 3.0),
+            self.pos.y - self.size * f32::sin(top_angle + 2.0 * std::f32::consts::PI / 3.0)
+        );
+        
+        // Bottom left vertex (330 degrees)
+        let v3 = Vec2::new(
+            self.pos.x + self.size * f32::cos(top_angle + 4.0 * std::f32::consts::PI / 3.0),
+            self.pos.y - self.size * f32::sin(top_angle + 4.0 * std::f32::consts::PI / 3.0)
+        );
+        
+        return DrawCall::Triangle(v1, v2, v3, self.color)
+    }
+
+    fn should_emit(&self) -> bool{
+        if let Ok(state) = self.machine.get_state().try_lock(){
+            if state.eq(&StateType::Hit){
+                return true
+            }
+        }
+        
+        return false
     }
 }
 
 impl GameEntity for Triangle{
+    #[inline(always)]
     fn get_id(&self) -> u64 {
         return self.id
     }
-    
+
     fn get_size(&self) -> f32 {
-        todo!()
+        return self.size
     }
-    
-    fn collides(&self,other: &dyn Collider) -> bool {
-        todo!()
+
+    fn collides(&self, other: &dyn Collider) -> bool {
+        return self.collider.collides_with(other)
     }
 
     fn get_collider(&self) -> Box<&dyn Collider> {
-        todo!()
+        return Box::new(&self.collider)
+    }
+}
+
+#[async_trait]
+impl Enemy for Triangle{
+    async fn new(id: u64, pos: Vec2, size: f32, color: Color, player_pos: Vec2, sender:Sender<Event>) -> Self where Self: Sized {
+        let enemy =  Triangle {
+            id: id,
+            pos: pos, 
+            size: size, 
+            speed: 100.0,
+            color: color,
+            target: player_pos,
+
+            sender: sender,
+            collider: CircleCollider::new(pos.x, pos.y, size),
+            machine: StateMachine::new(),
+
+            is_alive: true,
+            
+            emittion_configs: vec![(StateType::Hit, ConfigType::EnemyDeath)]
+        };
+
+        enemy.publish(Event::new((enemy.get_id(), enemy.emittion_configs.clone()), EventType::RegisterEmitterConf)).await;
+
+        return enemy
+    }
+
+    fn set_pos(&mut self, new_pos: Vec2){
+        self.pos = new_pos
+    }
+
+    fn is_alive(&self) -> bool{
+        return self.is_alive
+    }
+    
+    fn set_alive(&mut self, alive: bool) {
+        self.is_alive = alive;
+    }
+
+    fn force_state(&mut self, state: StateType){
+        self.machine.transition(state);
+    }
+
+    fn get_state(&self) -> Option<StateType>{
+        if let Ok(entry) = self.machine.get_state().try_lock(){
+            return Some(*entry)
+        }
+        return None
+    }
+
+}
+
+#[async_trait]
+impl Publisher for Triangle{
+    async fn publish(&self, event: Event){
+        let _ = self.sender.send(event);
     }
 }
 
 
+
 impl std::fmt::Debug for Triangle{
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Triangle")
+        f.debug_struct("Circle")
             .field("id", &self.id)
             .field("pos", &self.pos)
             .finish()