+++ src/entity_handler/enemy_type.rs
@@ -3,7 +3,6 @@
 #[derive(Clone, Copy)]
 pub enum EnemyType{
     Circle,
-    Ellipse,
     Triangle,
     Rect,
     Hexagon